Correct reference page
In particular:

- Remove superfluous contents entries
- Adjust desc indentation

The superfluous content lines were causing `pkgdown` not to be able to build the reference page, since the contents entries were judged to have no/non-character content., thereby breaking build process.

The desc indentation looks to be simply aesthetic in the code, so that these entries match others.